+++ lib/langchain/vectorsearch/qdrant.rb
@@ -5,10 +5,11 @@ module Langchain::Vectorsearch
     #
     # Wrapper around Qdrant
     #
-    # Gem requirements: gem "qdrant-ruby", "~> 0.9.3"
+    # Gem requirements:
+    #     gem "qdrant-ruby", "~> 0.9.3"
     #
     # Usage:
-    # qdrant = Langchain::Vectorsearch::Qdrant.new(url:, api_key:, index_name:, llm:, llm_api_key:)
+    #     qdrant = Langchain::Vectorsearch::Qdrant.new(url:, api_key:, index_name:, llm:)
     #
 
     # Initialize the Qdrant client
@@ -82,7 +83,7 @@ module Langchain::Vectorsearch
         collection_name: index_name,
         vectors: {
           distance: DEFAULT_METRIC.capitalize,
-          size: default_dimension
+          size: llm.default_dimension
         }
       )
     end